backend/requirements.txt
fastapi
uvicorn
pydantic-settings
pytest
httpx
sqlalchemy
alembic
psycopg[binary]
redis
ruff
